fix: Horas de enrolments tipo Double

+++ src/main/java/acme/features/student/enrolment/StudentEnrolmentShowService.java
@@ -87,7 +87,7 @@ public class StudentEnrolmentShowService extends AbstractService<Student, Enrolm
 		final Collection<Activity> activities = this.repository.findActivitiesByEnrolmentId(enrolmentId);
 		for (final Activity activity : activities) {
 			final long activityTime = Math.abs(activity.getEndDate().getTime() - activity.getStartDate().getTime());
-			final double activityTimeInHours = activityTime / 3600000;
+			final double activityTimeInHours = (double) activityTime / 3600000;
 			res += activityTimeInHours;
 		}
 		return res;
